test(size limit middleware): Implement test for SizeLimit middleware
